+++ include/supernote.h
@@ -24,7 +24,7 @@
 
 // Application info
 #define APP_NAME        L"OpenNote"
-#define APP_VERSION     L"0.6.0"
+#define APP_VERSION     L"0.6.1"
 #define APP_CLASS_NAME  L"OpenNoteMainWindow"
 
 // Limits